+++ tests/commands/export/test_export_causatives_verified_cmd.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 
 from scout.commands import cli
+from scout.constants.managed_variant import MANAGED_VARIANTS_INFILE_HEADER
 from scout.server.extensions import store
 
 
-def test_export_causatives(mock_app, case_obj):
+def test_export_causatives(mock_app, institute_obj, case_obj, user_obj):
     """Test the CLI command that exports causatives into vcf format"""
 
     runner = mock_app.test_cli_runner()
@@ -17,58 +18,86 @@ def test_export_causatives(mock_app, case_obj):
     runner.invoke(cli, ["load", "variants", case_obj["_id"], "--snv"])
     assert sum(1 for _ in store.variant_collection.find()) > 0
 
-    # update case registering a causative variant
+    ## WHEN marking a variant as causative
     variant_obj = store.variant_collection.find_one()
-    store.case_collection.find_one_and_update(
-        {"_id": case_obj["_id"]}, {"$set": {"causatives": [variant_obj["_id"]]}}
+    link = "markCausativelink"
+    store.mark_causative(
+        institute=institute_obj,
+        case=case_obj,
+        user=user_obj,
+        link=link,
+        variant=variant_obj,
     )
-    res = store.case_collection.find({"causatives": {"$exists": True}})
-    assert sum(1 for _ in res) == 1
+    assert store.event_collection.find(
+        {
+            "institute": "cust000",
+            "verb": {"$in": ["mark_causative", "mark_partial_causative"]},
+            "category": "case",
+        }
+    )
 
-    # Test the cli by not providing any options or arguments
-    result = runner.invoke(cli, ["export", "variants"])
+    # Test the CLI by not providing any options or arguments
+    result = runner.invoke(cli, ["export", "causatives"])
     assert result.exit_code == 0
     assert "#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO" in result.output
     # variant should be returned
     assert str(variant_obj["position"]) in result.output
 
-    # Test the cli by providing wrong collaborator
-    result = runner.invoke(cli, ["export", "variants", "-c", "cust666"])
+    # Test the CLI by providing wrong collaborator
+    result = runner.invoke(cli, ["export", "causatives", "-c", "cust666"])
     assert result.exit_code == 0
     assert "#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO" in result.output
     # variant should NOT be returned
     assert str(variant_obj["position"]) not in result.output
 
-    # Test the cli by providing the right collaborator, build and variant type
+    # Test the CLI by providing the right collaborator, build and variant type
     result = runner.invoke(
-        cli, ["export", "variants", "-c", case_obj["owner"], "--build", "37", "--category", "snv"]
+        cli, ["export", "causatives", "-c", case_obj["owner"], "--build", "37", "--category", "snv"]
     )
     assert result.exit_code == 0
     assert "#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO" in result.output
     # variant should be returned
     assert str(variant_obj["position"]) in result.output
 
-    # Test the cli by providing the document_id of the variant
-    result = runner.invoke(cli, ["export", "variants", "-d", variant_obj["document_id"]])
+    # Test the CLI by providing the document_id of the variant
+    result = runner.invoke(cli, ["export", "causatives", "-d", variant_obj["document_id"]])
     assert result.exit_code == 0
     assert "#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO" in result.output
     # variant should be returned
     assert str(variant_obj["position"]) in result.output
 
-    # Test the cli by providing the case_id of the variant
-    result = runner.invoke(cli, ["export", "variants", "--case-id", case_obj["_id"]])
+    # Test the CLI by providing the case_id of the variant
+    result = runner.invoke(cli, ["export", "causatives", "--case-id", case_obj["_id"]])
     assert result.exit_code == 0
     assert "#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO" in result.output
     # variant should be returned
     assert str(variant_obj["position"]) in result.output
 
-    # Test the cli by providing the case_id of the variantand and json option
-    result = runner.invoke(cli, ["export", "variants", "--case-id", case_obj["_id"], "--json"])
+    # Test the CLI by providing the case_id of the variant and and json option
+    result = runner.invoke(cli, ["export", "causatives", "--case-id", case_obj["_id"], "--json"])
     assert result.exit_code == 0
     assert "#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO" not in result.output
     # variant should be returned
     assert str(variant_obj["position"]) in result.output
     assert '"position": {}'.format(variant_obj["position"]) in result.output
+
+    # Test the CLI option to export causatives to a managed variants infile with time frame
+    result = runner.invoke(
+        cli,
+        [
+            "export",
+            "causatives",
+            "--as-managed",
+            "--managed-link-base-url",
+            "fakeurl",
+            "--within-days",
+            5,
+        ],
+    )
+    assert result.exit_code == 0
+    # variant should be returned
+    assert MANAGED_VARIANTS_INFILE_HEADER in result.output
+    assert str(variant_obj["position"]) in result.output
 
 
 def test_export_verified(mock_app, case_obj, user_obj, institute_obj):