+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [unreleased]
+### Added
+- Export causatives to managed variants infile via CLI, filtering variants by date (#6290)
 ### Fixed
 - Compress demo case rnafusion VCF and add an index (#6292)